@@ -1,56 +1,66 @@
 {
-  "name": "sdssample",
-  "version": "0.0.1",
-  "license": "Apache2.0",
+  "name": "sds-angular",
+  "description": "A sample that uses Sequential Data Store in an Angular-based environment.",
+  "version": "1.2.0",
+  "private": true,
+  "author": "OSIsoft, LLC",
+  "license": "Apache-2.0",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/osisoft/OSI-Samples-OCS.git",
+    "directory": "basic_samples/SDS/JavaScript/Angular"
+  },
   "scripts": {
     "ng": "ng",
-    "prestart": "echo 'WARNING: The web server used in this sample is intended for use in testing or debugging sample applications locally. It has not been reviewed for security issues.'",
+    "prestart": "echo WARNING: The web server used in this sample is intended for use in testing or debugging sample applications locally. It has not been reviewed for security issues.",
     "start": "ng serve",
     "build": "ng build",
     "test": "ng test",
     "lint": "ng lint",
-    "e2e": "ng e2e"
+    "e2e": "ng e2e --webdriver-update=false"
   },
-  "private": true,
   "dependencies": {
-    "@angular/animations": "8.2.12",
-    "@angular/cdk": "8.2.3",
-    "@angular/common": "8.2.12",
-    "@angular/compiler": "8.2.12",
-    "@angular/core": "8.2.12",
-    "@angular/forms": "8.2.12",
-    "@angular/material": "8.2.3",
-    "@angular/platform-browser": "8.2.12",
-    "@angular/platform-browser-dynamic": "8.2.12",
-    "@angular/router": "8.2.12",
-    "@ng-bootstrap/ng-bootstrap": "4.1.0",
+    "@angular/animations": "9.1.3",
+    "@angular/cdk": "9.2.1",
+    "@angular/common": "9.1.3",
+    "@angular/compiler": "9.1.3",
+    "@angular/core": "9.1.3",
+    "@angular/forms": "9.1.3",
+    "@angular/localize": "9.1.3",
+    "@angular/material": "9.2.1",
+    "@angular/platform-browser": "9.1.3",
+    "@angular/platform-browser-dynamic": "9.1.3",
+    "@angular/router": "9.1.3",
+    "@ng-bootstrap/ng-bootstrap": "6.1.0",
     "bluebird": "3.5.3",
     "core-js": "2.6.5",
     "moment": "2.24.0",
     "oidc-client": "1.7.0",
     "rxjs": "6.5.3",
     "rxjs-compat": "6.4.0",
-    "zone.js": "0.9.1"
+    "tslib": "^1.11.1",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.803.15",
-    "@angular/cli": "8.3.15",
-    "@angular/compiler-cli": "8.2.12",
-    "@angular/language-service": "8.2.12",
+    "@angular-devkit/build-angular": "~0.901.3",
+    "@angular/cli": "9.1.3",
+    "@angular/compiler-cli": "9.1.3",
+    "@angular/language-service": "9.1.3",
     "@types/jasmine": "3.3.9",
-    "@types/node": "11.11.3",
-    "codelyzer": "5.0.1",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
     "jasmine-core": "3.3.0",
+    "jasmine-reporters": "^2.3.2",
     "jasmine-spec-reporter": "4.2.1",
-    "karma": "4.0.1",
+    "karma": "^5.0.2",
     "karma-chrome-launcher": "2.2.0",
     "karma-cli": "2.0.0",
-    "karma-coverage-istanbul-reporter": "2.0.5",
+    "karma-coverage-istanbul-reporter": "2.1.1",
     "karma-jasmine": "2.0.1",
     "karma-jasmine-html-reporter": "1.4.0",
-    "protractor": "~5.4.0",
+    "protractor": "~5.4.4",
     "ts-node": "8.0.3",
-    "tslint": "5.14.0",
-    "typescript": "3.5.3"
+    "tslint": "6.1.1",
+    "typescript": "3.8.3"
   }
 }
