@@ -1,43 +1,29 @@
-// app.module.ts
-//
-
-import { BrowserModule } from '@angular/platform-browser';
+import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
 import { NgModule } from '@angular/core';
+import { BrowserModule } from '@angular/platform-browser';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
+
 import { AppComponent } from './app.component';
 import { SdsRestService } from './sds/sds.rest.service';
 import { DatasrcComponent } from './datasrc/datasrc.component';
-import { routing, appRoutingProviders  } from './app.routing';
-import {HttpClientModule, HTTP_INTERCEPTORS} from "@angular/common/http";
-import { OidcService, OAuthLoginComponent, OAuthLogoutComponent, OAuthCallbackHandlerGuard } from '../app/libraries/auth/ocs-auth';
+import { routing, appRoutingProviders } from './app.routing';
+import { OidcModule } from '../app/libraries/auth/ocs-auth';
 import { AuthenticationGuard } from '../app/libraries/auth/extra/authguard';
 import { AuthInterceptor } from '../app/libraries/auth/extra/authentication.interceptor';
 
 @NgModule({
-  declarations: [
-    AppComponent,
-    DatasrcComponent,
-    OAuthLoginComponent,
-    OAuthLogoutComponent
-  ],
-  imports: [
-    BrowserModule,
-    NgbModule.forRoot(),
-    routing,
-    HttpClientModule
-  ],
+  declarations: [AppComponent, DatasrcComponent],
+  imports: [BrowserModule, NgbModule, routing, OidcModule, HttpClientModule],
   providers: [
     appRoutingProviders,
     SdsRestService,
-    OidcService,
     AuthenticationGuard,
-    OAuthCallbackHandlerGuard,
     {
       provide: HTTP_INTERCEPTORS,
       useClass: AuthInterceptor,
-      multi: true
-    }
+      multi: true,
+    },
   ],
-  bootstrap: [AppComponent]
+  bootstrap: [AppComponent],
 })
-export class AppModule { }
+export class AppModule {}
