@@ -1,9 +1,9 @@
 ﻿using System;
+using System.Globalization;
 using System.IO;
 using System.Net.Http;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Configuration;
-using static System.Int32;
 
 namespace AuthorizationCodeFlow
 {
@@ -11,28 +11,28 @@ namespace AuthorizationCodeFlow
     {
         private static IConfiguration _configuration;
 
-        public static void Main(string[] args)
+        public static void Main()
         {
             bool test = false;
             try
             {
                 InitConfig();
 
-                if (SystemBrowser.openBrowser == null)
+                if (SystemBrowser.OpenBrowser == null)
                 {
-                    SystemBrowser.openBrowser = new OpenSystemBrowser();
+                    SystemBrowser.OpenBrowser = new OpenSystemBrowser();
                 }
                 else
                 {
                     test = true;
-                    SystemBrowser.password = GetConfigValue("Password");
-                    SystemBrowser.userName = GetConfigValue("UserName");
+                    SystemBrowser.Password = GetConfigValue("Password");
+                    SystemBrowser.UserName = GetConfigValue("UserName");
                 }
 
-                AuthorizationCodeFlow.OcsUrl = GetConfigValue("Resource");
-                AuthorizationCodeFlow.RedirectHost = GetConfigValue("RedirectHost");
-                AuthorizationCodeFlow.RedirectPort = Parse(GetConfigValue("RedirectPort"));
-                AuthorizationCodeFlow.RedirectPath = GetConfigValue("RedirectPath");
+                AuthorizationCode.OcsAddress = GetConfigValue("Resource");
+                AuthorizationCode.RedirectHost = GetConfigValue("RedirectHost");
+                AuthorizationCode.RedirectPort = int.Parse(GetConfigValue("RedirectPort"), CultureInfo.InvariantCulture);
+                AuthorizationCode.RedirectPath = GetConfigValue("RedirectPath");
 
                 var tenantId = GetConfigValue("TenantId");
                 var clientId = GetConfigValue("ClientId");
@@ -41,11 +41,11 @@ namespace AuthorizationCodeFlow
 
                 // Get access token.
                 var (accessToken, expiration) =
-                    AuthorizationCodeFlow.GetAuthorizationCodeFlowAccessToken(clientId, tenantId);
+                    AuthorizationCode.GetAuthorizationCodeFlowAccessToken(clientId, tenantId);
                 Console.WriteLine("Access Token: " + accessToken);
                 Console.WriteLine("Expires: " + expiration);
 
-                //  Make a request to Get Users endpoint
+                // Make a request to Get Users endpoint
                 var result1 = GetRequest($"{ocsUrl}/api/{apiVersion}/Tenants/{tenantId}/Users", accessToken).Result;
                 Console.WriteLine(result1
                     ? "Request succeeded"
@@ -57,7 +57,7 @@ namespace AuthorizationCodeFlow
             {
                 Console.WriteLine(e.ToString());
                 if (test)
-                    throw e;
+                    throw;
             }
 
             if (!test)
@@ -66,8 +66,8 @@ namespace AuthorizationCodeFlow
 
         private static async Task<bool> GetRequest(string endpoint, string accessToken)
         {
-            Console.WriteLine("Make request: ");
-            var request = new HttpRequestMessage()
+            Console.WriteLine("Make request:");
+            using var request = new HttpRequestMessage()
             {
                 Method = HttpMethod.Get,
                 RequestUri = new Uri(endpoint),
@@ -79,17 +79,14 @@ namespace AuthorizationCodeFlow
 
             try
             {
-                using (var client = new HttpClient())
-                {
-                    var response = await client.SendAsync(request);
-                    response.EnsureSuccessStatusCode();
+                using var client = new HttpClient();
+                var response = await client.SendAsync(request).ConfigureAwait(false);
+                response.EnsureSuccessStatusCode();
 
-                    // Uncomment this line to get the results of the calls
-                    //var responseBodyJson = JsonConvert.SerializeObject(response.Content.ReadAsStringAsync().Result, Formatting.Indented);
-                    //Console.WriteLine(responseBodyJson);
-
-                    return true;
-                }
+                // Uncomment this line to get the results of the calls
+                // var responseBodyJson = JsonConvert.SerializeObject(response.Content.ReadAsStringAsync().Result, Formatting.Indented);
+                // Console.WriteLine(responseBodyJson);
+                return true;
             }
             catch (HttpRequestException)
             {
@@ -105,15 +102,15 @@ namespace AuthorizationCodeFlow
                     .AddJsonFile("appsettings.json", optional: false, reloadOnChange: false)
                     .Build();
             }
-            catch (FileNotFoundException e)
+            catch (FileNotFoundException ex)
             {
-                Console.WriteLine("Config file missing");
-                throw e;
+                Console.WriteLine("Config file missing: " + ex);
+                throw;
             }
             catch (Exception ex)
             {
                 Console.WriteLine("Error while initiating configuration: " + ex.ToString());
-                throw ex;
+                throw;
             }
         }
 
@@ -137,10 +134,10 @@ namespace AuthorizationCodeFlow
 
                 return value;
             }
-            catch (Exception ex)
+            catch (Exception)
             {
                 Console.WriteLine($"Configuration issue");
-                throw ex;
+                throw;
             }
         }
     }
