@@ -6,27 +6,32 @@ namespace HybridFlow
 {
     public class OpenSystemBrowser : IOpenBrowser
     {
-        public void OpenBrowser(string url, string userName, string password)
+        public void OpenBrowser(string address, string userName, string password)
         {
+            if (string.IsNullOrEmpty(address))
+            {
+                throw new ArgumentException("Address cannot be null or empty", nameof(address));
+            }
+
             try
             {
-                Process.Start(url);
+                Process.Start(address);
             }
             catch
             {
                 // hack because of this: https://github.com/dotnet/corefx/issues/10361
                 if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
                 {
-                    url = url.Replace("&", "^&");
-                    Process.Start(new ProcessStartInfo("cmd", $"/c start {url}") { CreateNoWindow = true });
+                    address = address.Replace("&", "^&", StringComparison.OrdinalIgnoreCase);
+                    Process.Start(new ProcessStartInfo("cmd", $"/c start {address}") { CreateNoWindow = true });
                 }
                 else if (RuntimeInformation.IsOSPlatform(OSPlatform.Linux))
                 {
-                    Process.Start("xdg-open", url);
+                    Process.Start("xdg-open", address);
                 }
                 else if (RuntimeInformation.IsOSPlatform(OSPlatform.OSX))
                 {
-                    Process.Start("open", url);
+                    Process.Start("open", address);
                 }
                 else
                 {
