@@ -1,24 +1,14 @@
-using IdentityModel.OidcClient.Browser;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Http;
-using System;
-using System.IO;
+﻿using System;
 using System.Net;
 using System.Net.Sockets;
-using System.Text;
-using System.Threading;
 using System.Threading.Tasks;
+using IdentityModel.OidcClient.Browser;
 
 namespace HybridFlow
 {
     public class SystemBrowser : IBrowser
     {
-        public int Port { get; }
         private readonly string _path;
-        public static string userName;
-        public static string password;
-        public static IOpenBrowser openBrowser;
 
         public SystemBrowser(int? port = null, string path = null)
         {
@@ -34,137 +24,45 @@ namespace HybridFlow
             }
         }
 
+        public static string UserName { get; set; }
+        public static string Password { get; set; }
+        public static IOpenBrowser OpenBrowser { get; set; }
+
+        public int Port { get; }
+
+        public async Task<BrowserResult> InvokeAsync(BrowserOptions options)
+        {
+            if (options == null)
+            {
+                throw new ArgumentException("Options cannot be null", nameof(options));
+            }
+
+            using var listener = new LoopbackHttpListener(Port, _path);
+            OpenBrowser.OpenBrowser(options.StartUrl, UserName, Password);
+
+            try
+            {
+                var result = await listener.WaitForCallbackAsync().ConfigureAwait(false);
+                if (string.IsNullOrWhiteSpace(result))
+                {
+                    return new BrowserResult { ResultType = BrowserResultType.UnknownError, Error = "Empty response." };
+                }
+
+                return new BrowserResult { Response = result, ResultType = BrowserResultType.Success };
+            }
+            catch (TaskCanceledException ex)
+            {
+                return new BrowserResult { ResultType = BrowserResultType.Timeout, Error = ex.Message };
+            }
+        }
+
         private int GetRandomUnusedPort()
         {
             var listener = new TcpListener(IPAddress.Loopback, 0);
             listener.Start();
-            var port = ((IPEndPoint) listener.LocalEndpoint).Port;
+            var port = ((IPEndPoint)listener.LocalEndpoint).Port;
             listener.Stop();
             return port;
-        }
-
-        public async Task<BrowserResult> InvokeAsync(BrowserOptions options)
-        {
-            using (var listener = new LoopbackHttpListener(Port, _path))
-            {
-                openBrowser.OpenBrowser(options.StartUrl, userName, password);
-
-                try
-                {
-                    var result = await listener.WaitForCallbackAsync();
-                    if (string.IsNullOrWhiteSpace(result))
-                    {
-                        return new BrowserResult
-                            {ResultType = BrowserResultType.UnknownError, Error = "Empty response."};
-                    }
-
-                    return new BrowserResult {Response = result, ResultType = BrowserResultType.Success};
-                }
-                catch (TaskCanceledException ex)
-                {
-                    return new BrowserResult {ResultType = BrowserResultType.Timeout, Error = ex.Message};
-                }
-                catch (Exception ex)
-                {
-                    return new BrowserResult {ResultType = BrowserResultType.UnknownError, Error = ex.Message};
-                }
-            }
-        }
-    }
-
-    public class LoopbackHttpListener : IDisposable
-    {
-        const int DefaultTimeout = 60 * 5; // 5 mins (in seconds)
-
-        IWebHost _host;
-        TaskCompletionSource<string> _source = new TaskCompletionSource<string>();
-
-        public string Url { get; }
-
-        public LoopbackHttpListener(int port, string path = null)
-        {
-            path = path ?? string.Empty;
-            if (path.StartsWith("/")) path = path.Substring(1);
-
-            Url = $"https://127.0.0.1:{port}/{path}";
-
-            _host = new WebHostBuilder()
-                .UseKestrel()
-                .UseUrls(Url)
-                .Configure(Configure)
-                .Build();
-            _host.Start();
-        }
-
-        public void Dispose()
-        {
-            Task.Run(async () =>
-            {
-                await Task.Delay(500);
-                _host.Dispose();
-            });
-        }
-
-        void Configure(IApplicationBuilder app)
-        {
-            app.Run(async ctx =>
-            {
-                if (ctx.Request.Method == "GET")
-                {
-                    SetResult(ctx.Request.QueryString.Value, ctx);
-                }
-                else if (ctx.Request.Method == "POST")
-                {
-                    if (!ctx.Request.ContentType.Equals("application/x-www-form-urlencoded",
-                        StringComparison.OrdinalIgnoreCase))
-                    {
-                        ctx.Response.StatusCode = 415;
-                    }
-                    else
-                    {
-                        using (var sr = new StreamReader(ctx.Request.Body, Encoding.UTF8))
-                        {
-                            var body = await sr.ReadToEndAsync();
-                            SetResult(body, ctx);
-                        }
-                    }
-                }
-                else
-                {
-                    ctx.Response.StatusCode = 405;
-                }
-            });
-        }
-
-        private void SetResult(string value, HttpContext ctx)
-        {
-            try
-            {
-                ctx.Response.StatusCode = 200;
-                ctx.Response.ContentType = "text/html";
-                ctx.Response.WriteAsync("<h1>You can now return to the application.</h1>");
-                ctx.Response.Body.Flush();
-
-                _source.TrySetResult(value);
-            }
-            catch
-            {
-                ctx.Response.StatusCode = 400;
-                ctx.Response.ContentType = "text/html";
-                ctx.Response.WriteAsync("<h1>Invalid request.</h1>");
-                ctx.Response.Body.Flush();
-            }
-        }
-
-        public Task<string> WaitForCallbackAsync(int timeoutInSeconds = DefaultTimeout)
-        {
-            Task.Run(async() =>
-            {
-                await Task.Delay(timeoutInSeconds * 1000);
-                _source.TrySetCanceled();
-            });
-
-            return _source.Task;
         }
     }
 }
