@@ -1,12 +1,4 @@
-// <copyright file="SdsInterpolationMode.cs" company="OSIsoft, LLC">
-//
-// </copyright>
-
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace SdsRestApiCore
+﻿namespace SdsRestApiCore
 {
     public enum SdsInterpolationMode
     {
@@ -14,6 +6,6 @@ namespace SdsRestApiCore
         Continuous = 0,
         StepwiseContinuousLeading = 1,
         StepwiseContinuousTrailing = 2,
-        Discrete = 3
+        Discrete = 3,
     }
 }
