@@ -1,31 +1,11 @@
-// <copyright file="SdsStreamViewProperty.cs" company="OSIsoft, LLC">
-//
-// </copyright>
-
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace SdsRestApiCore
+﻿namespace SdsRestApiCore
 {
     public class SdsStreamViewProperty
     {
-        public string SourceId
-        {
-            get;
-            set;
-        }
+        public string SourceId { get; set; }
 
-        public string TargetId
-        {
-            get;
-            set;
-        }
+        public string TargetId { get; set; }
 
-        public SdsStreamView SdsStreamView
-        {
-            get;
-            set;
-        }
+        public SdsStreamView SdsStreamView { get; set; }
     }
 }
