@@ -1,19 +1,7 @@
-// <copyright file="SdsTenant.cs" company="OSIsoft, LLC">
-//
-// </copyright>
-
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace SdsRestApiCore
+﻿namespace SdsRestApiCore
 {
     public class SdsTenant
     {
-        public string Id
-        {
-            get;
-            set;
-        }
+        public string Id { get; set; }
     }
 }
