@@ -1,18 +1,16 @@
 // Protractor configuration file, see link for more information
 // https://github.com/angular/protractor/blob/master/lib/config.ts
 
-const { SpecReporter } = require('jasmine-spec-reporter');
+const { JUnitXmlReporter } = require('jasmine-reporters');
 
 exports.config = {
   allScriptsTimeout: 110000,
-  specs: [
-    './src/**/*.e2e-spec.ts'
-  ],
+  specs: ['./src/**/*.e2e-spec.ts'],
   capabilities: {
-      'browserName': 'ChromeHeadless',
-      chromeOptions: {
-          args: ["--headless", "--disable-gpu", "--window-size=800x600"]
-      }
+    browserName: 'chrome',
+    'goog:chromeOptions': {
+      args: ['--headless', '--no-sandbox'],
+    },
   },
   directConnect: true,
   baseUrl: 'http://localhost:4200/',
@@ -20,12 +18,14 @@ exports.config = {
   jasmineNodeOpts: {
     showColors: true,
     defaultTimeoutInterval: 3000000,
-    print: function() {}
+    print: function () {},
   },
   onPrepare() {
     require('ts-node').register({
-      project: require('path').join(__dirname, './tsconfig.e2e.json')
+      project: require('path').join(__dirname, './tsconfig.e2e.json'),
     });
-    jasmine.getEnv().addReporter(new SpecReporter({ spec: { displayStacktrace: true } }));
-  }
+    jasmine
+      .getEnv()
+      .addReporter(new JUnitXmlReporter({ savePath: 'TestResults' }));
+  },
 };
