@@ -1,6 +1,3 @@
-// environment.prod.ts
-//
-
 export const environment = {
-  production: true
+  production: true,
 };
