@@ -1,12 +1,17 @@
+"""This script tests the SDS Time Series Python sample script"""
+
 import unittest
 from .program import main
 
 
-class ProgramTestCase(unittest.TestCase):
+class SDSTimeSeriesPythonSampleTests(unittest.TestCase):
+    """Tests for the SDS Time Series Python sample"""
 
-    def test_itRuns(self):
-        main()
+    @classmethod
+    def test_main(cls):
+        """Tests the SDS Time Series Python main sample script"""
+        main(True)
 
 
 if __name__ == "__main__":
-    unittest.main(True)
+    unittest.main()
